Add login/logout functionality and restrict pages

+++ base/views.py
@@ -1,16 +1,42 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 
 from .models import Room, Topic, Message, User, BlogPost
 from .forms import RoomForm, PostForm
 
-def login(request):
-    pass
+def loginUser(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
 
-def logout(request):
-    pass
+    if request.method == 'POST':
+        email = request.POST.get('email')
+        password = request.POST.get('password')
 
-def register(request):
+        try:
+            user = User.objects.get(email=email)
+        except:
+            messages.error(request, "User does not exist")
+        else:
+            user = authenticate(request, email=email, password=password)
+
+            if user is not None:
+                login(request, user)
+                return redirect('home')
+            else:
+                messages.error(request, "Username OR password incorrect")
+    
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerUser(request):
     pass
 
 def userProfile(request, pk):
@@ -38,6 +64,7 @@ def room(request, pk):
     context = {"room": room, "room_messages": room_messages, "participants": participants}
     return render(request, "base/room.html", context)
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -49,9 +76,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return redirect('home')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -60,8 +92,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return redirect('home')
+    
     if request.method == "POST":
         room.delete()
         return redirect('home')
